refactor: consolidate duplicate test helpers into lib (#130)
