+++ hooks/ObligationStateMachines/HookDocTracker/HookDocStateMachine.test.ts
@@ -22,7 +22,7 @@ import {
   isSilentNoOp,
   buildStopInput as makeStopInput,
   buildToolInput as makeToolInput,
-} from "@hooks/hooks/ObligationStateMachines/test-helpers";
+} from "@hooks/lib/test-helpers";
 import type { ObligationDeps } from "@hooks/lib/obligation-machine";
 
 // ─── Helpers ──────────────────────────────────────────────────────────────────
