+++ hooks/ObligationStateMachines/TestObligationTracker/TestObligationStateMachine.test.ts
@@ -13,7 +13,7 @@ import {
   isSilentNoOp,
   buildStopInput as makeStopInput,
   buildToolInput as makeToolInput,
-} from "@hooks/hooks/ObligationStateMachines/test-helpers";
+} from "@hooks/lib/test-helpers";
 
 // ─── Helpers ──────────────────────────────────────────────────────────────────
