+++ package.json
@@ -17,7 +17,8 @@
     "@swc/core": "^1.15.21",
     "@types/bun": "^1.3.10",
     "husky": "^9.0.0",
-    "type-fest": "^5.5.0"
+    "type-fest": "^5.5.0",
+    "typescript": "^6.0.2"
   },
   "dependencies": {
     "@anthropic-ai/claude-agent-sdk": "^0.2.98",
